Fix bulleted list formatting and clarify license
GNU has multiple licenses, so "GNU License" is ambiguous. I changed it to "GNU GPLv3" because that is the license in the file LICENSE.md, but I would recommend changing it to "GNU GPLv3 or later" (or even better, "GNU GPLv2 or later" if you want to be compatible with GPLv2), so that if a GPLv4 license is released, Aspine will be compatible with projects that use GPLv4. I have not made this change because I do not hold the copyright to the software and do not have the legal right to make such a decision about the license of the software.