Added hosting instructions
I added instructions to host Aspine locally for testing purposes. I prepended
`node` to the command `./serve.js insecure` because `./serve.js` assumes that
the file `serve.js` has the execute bit enabled in Unix permissions, but
the repository may be cloned into a directory on a filesystem without Unix
permissions, such as NTFS (like on my computer), or the user may be running
a non-Unix-like operating system such as Windows (I'm not sure how Windows
handles this).