Add backup and restore functionality
The command-line options "json" and "out" have been introduced. If the "json"
option is given, the json file at the given path is used instead of scraping
Aspen. If the "out" option is given, Aspen is scraped as usual, but the data
are dumped to a json file at the given path.