Add /schedule endpoint and factor out schedule code
1) The schedule-scraping has been commented out of scrape_student in
scrape.js, instead exporting scrape_schedule directly.
2) A new endpoint, '/schedule', has been declared in serve.js, from which
the schedule can be retrieved as JSON (because it is no longer included
in the JSON file from /data).
3) The JavaScript in home.html now performs the schedule request separately
with a separate callback function, and clock.js no longer assumes
that tableData.schedule must exist if tableData exists.

A similar treatment will be given to other components that are scraped
until the loading has become sufficiently modular (lazy loading)
to make it feasible to retrieve data from past terms and to extend
Aspine in other ways in the future.