Correctly initialize quarter dropdown selector
After adding scrape_overview to scrape.js, it is now possible to collect
the overall grades from all classes and all quarters very quickly. Using
this information it is possible to correctly assign GPA's to each of the
options in the term dropdown selector after the first call to /data.