Fix GPA calculator by taking special cases into account.
Resolve: The GPAs change when the user clicks on them. Are the values for
all terms fetched from Aspen somehow and then recalculated as each term is loaded?

Forgot to take into account that even if a student has a grade over 100
in a class, as far as GPA is considered, the student has a 100 in the
class.