Stop hard-coding S1 and increase class threads
Aspine was hard-coded to work only with the first semester, looking for
classes in "FY" or "S1".
Because the second semester has now begun, and Aspine can fetch prior
terms, hard-coding the current term is no longer feasible.
A global variable TERM_NAMES has been added to store the possible "terms"
that correspond to each quarter, and this is now used for checking if
a class belongs to a certain quarter.

Additionally, CLASS_THREADS has been increased to 12 because some people
have more than 10 classes on their schedule.