Make build script for "lite" version of Aspine
We can run this build script whenever we tag a new version and
distribute the resulting client-side-only version of Aspine. This is
useful for people who want to view a JSON file from Aspine 2.x after
compatibility breaks.

This commit also adapts some other code to work with the lite version.