Generalize custom-select classes
Some of the classes relating to "custom-select" were used specifically
to refer to the GPA/term dropdown. They have now been renamed to include
"gpa" in their names and have been supplemented with generic equivalents.
This will make it easier to reuse CSS when another dropdown is added to
the top bar.