"Add assignment" button makes it add to current category (#162)
adds to buttonFuntions.js and home.js, changing the newAssignment function so it adds to the current category filter (and if none is selected just the first category filter), and tracks what the filter is currently. I did this based off a variable and not the filter because it's easier.