Fix colors in sidenav
Match the sidenav color scheme to the colors used in the navbar.

Change highlighting so that tabs turn a darker grey when hovered over
and an even darker grey when selected.

Make the dropdown menus take on a darker color when expanded (in order
to distinguish them from other buttons).

Co-authored-by: psvenk <45520974+psvenk@users.noreply.github.com>