Remove shebang lines from scrape/serve
Because the recommended way to start the Aspine server is `npm start`
and not `./serve.js` (especially in light of the oncoming move to the
new REST API scraper), there is no use in having shebang lines in
serve.js or scrape.js (particularly considering that the #!/bin/sh
shebang is hacky and may confuse some editors, causing them to treat
these files as shell scripts).

Also, this commit adds a semicolon to an import in serve.js.