Use day of week to determine schedule + B/S
The day of the week of the current time or the date_override is now used
to determine which schedule variant (Mon/Tue, Wed, Thu/Fri) is used as
well as whether it is a black day or a silver day.

Additionally, the black/silver slider on the schedule tab has been
removed.