Show score, release/due date on all assignments
Create a table for assignments without stats that shows your score, the
maximum score, date assigned, date due, and assignment feedback.
For now, I'm trying to get more data to show up in the modal to remove
table columns on mobile, but eventually, the info in the modal will be
transferred to a separate region under the row in the table.