Make adjustColumns trigger after data loaded
The load event listener creates a race condition which results in the
assignments table header not loading properly on mobile. Instead, call
adjustColumns just after the assignments and categories tables are
populated with data.

Additionally, this commit has adjustColumns take in as a parameter the
table on which it operates, enabling the resize event listener to be
initialized only once the tables are built. This prevents edge case race
conditions and makes the code more generalizable (if we want to use
adjustColumns for, e.g., the schedule table in the future).